refactor: move flow_context_resolver
