+++ src/promptflow/promptflow/core/_flow.py
@@ -243,7 +243,7 @@ class Flow(FlowBase):
 
     def invoke(self, inputs: dict) -> "LineResult":
         """Invoke a flow and get a LineResult object."""
-        from promptflow.core._flow_context_resolver import FlowContextResolver
+        from promptflow._sdk.entities._flow._flow_context_resolver import FlowContextResolver
 
         invoker = FlowContextResolver.resolve(flow=self)
         result = invoker._invoke(
@@ -411,7 +411,7 @@ class AsyncFlow(Flow):
 
     async def invoke_async(self, inputs: dict) -> "LineResult":
         """Invoke a flow and get a LineResult object."""
-        from promptflow.core._flow_context_resolver import FlowContextResolver
+        from promptflow._sdk.entities._flow._flow_context_resolver import FlowContextResolver
 
         invoker = FlowContextResolver.resolve_async_invoker(flow=self)
         result = await invoker._invoke_async(
