+++ src/promptflow/tests/sdk_cli_global_config_test/e2etests/test_global_config.py
@@ -4,8 +4,8 @@ import mock
 import pytest
 
 from promptflow._sdk._load_functions import load_flow
+from promptflow._sdk.entities._flow._flow_context_resolver import FlowContextResolver
 from promptflow._sdk.operations._local_azure_connection_operations import LocalAzureConnectionOperations
-from promptflow.core._flow_context_resolver import FlowContextResolver
 
 FLOWS_DIR = Path(__file__).parent.parent.parent / "test_configs" / "flows"
 DATAS_DIR = Path(__file__).parent.parent.parent / "test_configs" / "datas"
